+++ AddOns/Platynator/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Platynator
 
-## [325](https://github.com/TheMouseNest/Platynator/tree/325) (2026-02-19)
-[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/324...325) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
+## [326](https://github.com/TheMouseNest/Platynator/tree/326) (2026-02-20)
+[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/325...326) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
 
-- Classic: Fix channeled cast bars  
+- [Fixes #211] Retail: Fix soft targeting not updating when hard target changes  
+- Classic: Tag scatter shot as CC  
+- Retail: Fix "Defensive" filter not applying to buffs  
+- Fix "Friendly" auras showing on duelled players  
+- Hide "Defensive" option for classic, as the filter doesn't exist  
+- Add "Defensive" option to buffs widget